requirements.txt
flask
tensorflow
scikit-learn
numpy

opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask